requirements.txt
google-genai>=0.3.0
python-dotenv>=1.0.0
langchain>=0.0.300
langchain-google-genai>=0.0.1
langchain-community>=0.0.20
faiss-cpu
